rpi/requirements.txt
# --- Core API ---
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3

# --- Vision / ML (RPi 5) ---
opencv-python-headless==4.10.0.84
numpy==1.26.4
ultralytics==8.3.40

# --- I/O ---
pyserial==3.5
smbus2==0.4.3
RPi.GPIO==0.7.1; sys_platform == "linux"

# --- Audio / TTS ---
pyttsx3==2.98

# --- System metrics ---
psutil==6.1.0

# --- Dev / Test ---
pytest==8.3.4
httpx==0.28.1
